update to stump

+++ stump.py
@@ -258,7 +258,8 @@ if __name__ == '__main__':
         call("git clone --no-checkout {} {}".format(url, tmp), shell=True)
 
         # Copy the contents of the real directory
-        shutil.rmtree(prefix + '/.git')
+        if os.path.exists(prefix + '/.git'):
+            shutil.rmtree(prefix + '/.git')
         copy_tree(prefix, tmp, lambda d, files: ['.git'])
 
         call('git -C {0} add --all'.format(tmp), shell=True)
